[refactoring] split command buffer from handler file
