+++ SSD/SSD/command_buffer.cpp
@@ -1,0 +1,17 @@
+#include "command_buffer.h"
+
+CommandBuffer::CommandBuffer() {}
+
+string CommandBuffer::readDataOnBuffer(int lba) {
+	return "";
+}
+
+int CommandBuffer::getUsableBufferSize() {
+	return 0;
+}
+
+void CommandBuffer::insertCmd(int lba, int size) {
+}
+
+void CommandBuffer::insertCmd(int lba, string value) {
+}
